feat(dedicated.netapp): refactor efs network (#11016)
ref: MANAGER-12423
