+++ packages/manager/modules/netapp/src/dashboard/network-configuration/component.js
@@ -5,8 +5,7 @@ export default {
   bindings: {
     goBack: '<',
     trackClick: '<',
-    trackSuccess: '<',
-    trackError: '<',
+    trackPage: '<',
     vrackServices: '<',
     vracks: '<',
     createVrackServiceLink: '<',
