+++ packages/manager/modules/netapp/src/dashboard/network-configuration/routing.js
@@ -23,13 +23,9 @@ export default /* @ngInject */ ($stateProvider) => {
           type: 'action',
           name: `${TRACKING_BASE}::${tracker}`,
         }),
-      trackSuccess: /* @ngInject */ (atInternet) => () =>
+      trackPage: /* @ngInject */ (atInternet) => (tracker) =>
         atInternet.trackPage({
-          name: `${TRACKING_BASE}-success`,
-        }),
-      trackError: /* @ngInject */ (atInternet) => () =>
-        atInternet.trackPage({
-          name: `${TRACKING_BASE}-error`,
+          name: `${TRACKING_BASE}-${tracker}`,
         }),
       goBack: /* @ngInject */ ($state, trackClick) => (trackingAction) => {
         trackClick(trackingAction);
