+++ packages/manager/modules/netapp/src/dashboard/vrack-configuration/constants.js
@@ -1,10 +1,5 @@
-export const LABELS = {
-  VRACK: 'vRack',
-};
-
 export const TRACKING_BASE = 'netapp::dashboard::configure-vrack';
 
 export default {
-  LABELS,
   TRACKING_BASE,
 };
